+++ apollo-runtime-common/build.gradle.kts
@@ -32,6 +32,7 @@ kotlin {
     val jvmMain by getting {
       dependsOn(commonMain)
       dependencies {
+        api(groovy.util.Eval.x(project, "x.dep.okHttp.okHttp"))
       }
     }
