+++ apollo-runtime-kotlin/build.gradle.kts
@@ -25,6 +25,7 @@ kotlin {
     val commonMain by getting {
       dependencies {
         api(project(":apollo-api"))
+        api(project(":apollo-runtime-common"))
         api(groovy.util.Eval.x(project, "x.dep.okio"))
         api(groovy.util.Eval.x(project, "x.dep.uuid"))
         api(groovy.util.Eval.x(project, "x.dep.kotlin.coroutines"))
@@ -32,9 +33,8 @@ kotlin {
     }
 
     val jvmMain by getting {
-      dependsOn(commonMain)
       dependencies {
-        api(groovy.util.Eval.x(project, "x.dep.okHttp.okHttp4"))
+        api(groovy.util.Eval.x(project, "x.dep.okHttp.okHttp"))
       }
     }
