+++ gradle/dependencies.gradle
@@ -17,8 +17,7 @@ def versions = [
     moshi                 : '1.11.0',
     okHttp4               : '4.9.0',
     okHttp                : '3.12.11', // Keep this to keep supporting older Android devices
-    okio                  : '2.9.0',
-    okioSnapshot          : '2.10.0-SNAPSHOT',
+    okio                  : '2.10.0',
     rxjava                : '2.2.20',
     rxjava3               : '3.0.7',
     rxandroid             : '2.0.1',
@@ -77,7 +76,6 @@ ext.dep = [
         logging4      : "com.squareup.okhttp3:logging-interceptor:$versions.okHttp4",
     ],
     okio                  : "com.squareup.okio:okio-multiplatform:$versions.okio",
-    okioFiles          : "com.squareup.okio:okio-files:$versions.okioSnapshot",
     poet                  : [
         java  : "com.squareup:javapoet:$versions.javaPoet",
         kotlin: "com.squareup:kotlinpoet:$versions.kotlinPoet",
